temp fix for uninstal so install works durring dev

+++ administration/com_brazitrac/uninstall--COPY.brazitrac.php
@@ -9,10 +9,19 @@
 // Don't allow direct linking
 defined('_JEXEC') or die('Restricted Access');
 
-
+function com_uninstall()
+{
 	$database = JFactory::getDBO();
 	
-
+	// setup settings
+    require('components/com_brazitrac/classes/dbhelper.php');  
+    require('components/com_brazitrac/classes/factory.php');  
+    require('components/com_brazitrac/classes/config.php');
+	require('components/com_brazitrac/admin.brazitrac.html.php');
+	$btrac = WFactory::getConfig();
+	
+	if ( $btrac->get( 'upgrade' ) == 0 )
+	{
 		// uninstall
 		$database->setQuery( "DROP TABLE IF EXISTS `#__brazitrac_category`;" );
 		$database->query();
@@ -30,9 +39,9 @@ defined('_JEXEC') or die('Restricted Access');
 		$database->query();
 		$database->setQuery( "DROP TABLE IF EXISTS `#__brazitrac_users`;" );
 		$database->query();
-
+	}
 
 	return '<div style="text-align: center;"><h3>Brazitrac Ticket System</h3><p>Thanks for using the BraziTrac TicketSystem</p></div>';
-
+}
 
 ?>
